third_party/WebKit/LayoutTests/http/tests/inspector/elements/html-link-import.html

<html>
<head>
<link rel="import" href="../resources/imported-document.html">
<script src="../inspector-test.js"></script>
<script src="../elements-test.js"></script>
<script>

function test()
{
    // Warm up highlighter module.
    runtime.loadModulePromise("source_frame").then(function() {
        InspectorTest.expandElementsTree(callback);
    }).done();

    function callback()
    {
        InspectorTest.dumpElementsTree();
        InspectorTest.completeTest();
    }

}
</script>
</head>

<body onload="runTest()">

<p>This test verifies that imported document is rendered within the import link.</p>

</body>
</html>
