third_party/WebKit/LayoutTests/http/tests/inspector/elements/html-link-import-expected.txt

This test verifies that imported document is rendered within the import link.

- <html>
    - <head>
        - <link rel="import" href="../resources/imported-document.html">
            - #document
                - <html>
                    - <head>
                          <link rel="import" href="imported-document.html">
                        - <link rel="import" href="imported-document-2.html">
                            - #document
                                - <html>
                                      <head></head>
                                    - <body>
                                        - <div class="warning">
                                              <style scoped>\n    h3 {\n      color: red;\n    }\n  </style>
                                              <h3>Warning!</h3>
                                              <p>This page is under construction</p>
                                          </div>
                                        - <div class="outdated">
                                              <h3>Heads up!</h3>
                                              <p>This content may be out of date</p>
                                          </div>
                                      </body>
                                  </html>
                          </link>
                      </head>
                    - <body>
                          <div class="first-level">\n</div>
                      </body>
                  </html>
          </link>
          <script src="../inspector-test.js"></script>
          <script src="../elements-test.js"></script>
        - <script>
              \nfunction test()\n{\n    // Warm up highlighter module.\n    runtime.loadModulePromise("source_frame").then(function() {\n        InspectorTest.expandElementsTree(callback);\n    }).done();\n\n    function callback()\n    {\n        InspectorTest.dumpElementsTree();\n        InspectorTest.completeTest();\n    }\n\n}\n
          </script>
      </head>
    - <body onload="runTest()">
          <p>This test verifies that imported document is rendered within the import link.</p>
      </body>
  </html>
